+++ projects/core/src/cart/store/reducers/multi-cart.reducer.ts
@@ -3,6 +3,7 @@ import { LoaderAction } from '../../../state/utils/loader/loader.action';
 import { CartActions } from '../actions/index';
 
 export const activeCartInitialState = '';
+export const wishListinitialState = '';
 
 export function activeCartReducer(
   state = activeCartInitialState,
@@ -47,6 +48,18 @@ export function cartEntitiesReducer(
       return action.payload.cart;
     case CartActions.SET_FRESH_CART:
       return action.payload;
+  }
+  return state;
+}
+
+export function wishListReducer(
+  state = wishListinitialState,
+  action: CartActions.WishListActions
+): string {
+  switch (action.type) {
+    case CartActions.CREATE_WISH_LIST_SUCCESS:
+    case CartActions.LOAD_WISH_LIST_SUCCESS:
+      return action.meta.entityId as string;
   }
   return state;
 }
